fix sync-mode timing drift and audible tempo wobble against external clocks
Tracked external tempo was rounded to a whole BPM before driving playback,
leaving a small but systematic rate bias that compounded into growing
lateness over a long session; the loop's timing anchor was also rebased to
the current time on every tempo update, adding a forward-only phase jump
each time. Playback now tracks the continuous, unrounded tempo and only
ever changes rate, never phase, so it stays locked to the external clock
indefinitely.

+++ src/loop_engine/mod.rs
@@ -31,7 +31,7 @@ pub(crate) enum LoopCommand {
     SyncStart,
     SyncContinue,
     SyncStop,
-    SyncBpmUpdate(u32),
+    SyncBpmUpdate(f64),
 }
 
 pub struct LoopEngine {
@@ -142,7 +142,7 @@ impl LoopEngine {
         let _ = self.sender.send(LoopCommand::SyncStop);
     }
 
-    pub fn sync_bpm_update(&self, bpm: u32) {
+    pub fn sync_bpm_update(&self, bpm: f64) {
         let _ = self.sender.send(LoopCommand::SyncBpmUpdate(bpm));
     }
 }
@@ -2119,7 +2119,7 @@ mod tests {
         wait_for_state(&engine, EngineState::Running, 500);
         std::thread::sleep(Duration::from_millis(100));
 
-        engine.sync_bpm_update(150);
+        engine.sync_bpm_update(150.0);
 
         std::thread::sleep(Duration::from_millis(200));
         assert_eq!(
